// src/interfaces/IWeatherData.ts
export interface IWeatherData {
  sunrise?: number;
  sunset?: number;
  feelslike?: number;
  tempmax?: number;
  tempmin?: number;
  humidity?: number;
  pressure?: number;
  visibility?: number;
  windSpeed?: number;
  temprature?: number;
  city?: string;
  country?: string;
  description?: string;
  icon?: string;
}